Tests/LibWeb/Text/expected/css/insert-rule-keyframes-shadow-tree-invalidation.txt

opacity before keyframes: 1
opacity after keyframes: 0.25
